+++ src/components/InformationEdit/UserImg.js
@@ -2,9 +2,80 @@ import React, { useState, useRef } from "react";
 import { post, put, axios } from "axios";
 import Avatar from "@mui/material/Avatar";
 import styled from "styled-components";
+import {
+  Box,
+  Chip,
+  FormControl,
+  InputLabel,
+  MenuItem,
+  OutlinedInput,
+  Select,
+  TextField,
+  useTheme,
+} from "@mui/material";
+import { useParams } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
+import { searchAsync, updateAsync } from "../../store/user/user";
+import { useEffect } from "react";
+//-------------------------------------------------------------스택관련 함수
+const ITEM_HEIGHT = 48;
+const ITEM_PADDING_TOP = 8;
+const MenuProps = {
+  PaperProps: {
+    style: {
+      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
+      width: 250,
+    },
+  },
+};
+
+const names = [
+  "JAVA",
+  "JAVA_SCRIPT",
+  "REACT",
+  "VUE",
+  "PYTHON",
+  "SPRING",
+  "TYPE_SCRIPT",
+  "ANGULAR",
+  "CPP",
+  "KOTILIN",
+  "SVELTE",
+];
+
+function getStyles(name, personName, theme) {
+  return {
+    fontWeight:
+      personName.indexOf(name) === -1
+        ? theme.typography.fontWeightRegular
+        : theme.typography.fontWeightMedium,
+  };
+}
+//------------------------------------------------------------------
 // 내 정보 수정 화면에서 이미지 부분
 
 const UserImg = () => {
+  const { member_id } = useParams();
+  const dispatch = useDispatch();
+  const nickName = useSelector(state => state.user.nickName);
+  const userStack = useSelector(state => state.user.userStack);
+  const theme = useTheme();
+  const [personName, setPersonName] = useState(userStack);
+  const [updateNickName, setUpdateNickName] = useState(nickName);
+  const [password, setPassword] = useState("");
+  useEffect(() => {
+    dispatch(searchAsync(member_id));
+  }, []);
+  const handleChange = event => {
+    const {
+      target: { value },
+    } = event;
+    setPersonName(
+      // On autofill we get a stringified value.
+      typeof value === "string" ? value.split(",") : value
+    );
+  };
+  //--------------------스택관련 위 아래는 이미지관련
   const [Image, setImage] = useState({
     image_file: "",
     preview_URL:
@@ -81,6 +152,131 @@ const UserImg = () => {
           <ImageDelete onClick={deleteImage}>이미지 삭제</ImageDelete>
         </div>
       </ImgBox>
+      <UserNameWrapper>
+        <h3 style={{ margin: "30px" }}>닉네임</h3>
+        <TextField
+          margin="normal"
+          required
+          id="nickName"
+          label="닉네임"
+          value={updateNickName}
+          onChange={event => {
+            setUpdateNickName(event.target.value);
+          }}
+          autoFocus
+        />
+      </UserNameWrapper>
+      <p
+        style={{
+          color: "#868e96",
+          fontSize: ".875rem",
+          paddingTop: "15px",
+          paddingLeft: "30px",
+          paddingBottom: "15px",
+          borderBottom: "1px solid grey",
+        }}
+      >
+        Co-meet에서 사용되는 이름입니다.
+      </p>
+      <TechWrapper>
+        <h3>관심 기술 태그</h3>
+        <FormControl sx={{ m: 1, width: 300 }}>
+          <InputLabel id="demo-multiple-chip-label">TechStack</InputLabel>
+          <Select
+            labelId="demo-multiple-chip-label"
+            id="demo-multiple-chip"
+            multiple
+            value={personName}
+            onChange={handleChange}
+            input={<OutlinedInput id="select-multiple-chip" label="Chip" />}
+            renderValue={selected => (
+              <Box sx={{ display: "flex", flexWrap: "wrap", gap: 0.5 }}>
+                {selected.map(value => (
+                  <Chip key={value} label={value} />
+                ))}
+              </Box>
+            )}
+            MenuProps={MenuProps}
+          >
+            {names.map(name => (
+              <MenuItem
+                key={name}
+                value={name}
+                style={getStyles(name, personName, theme)}
+              >
+                {name}
+              </MenuItem>
+            ))}
+          </Select>
+        </FormControl>
+      </TechWrapper>
+      <p
+        style={{
+          color: "#868e96",
+          fontSize: ".875rem",
+          paddingTop: "15px",
+          paddingLeft: "30px",
+          paddingBottom: "15px",
+          borderBottom: "1px solid grey",
+        }}
+      >
+        관심 있는 기술 태그를 입력해주세요.
+      </p>
+      <UserPasswordWrapper>
+        <h3 style={{ margin: "30px" }}>비밀번호</h3>
+        <TextField
+          margin="normal"
+          required
+          id="password"
+          label="비밀번호"
+          value={password}
+          onChange={event => {
+            setPassword(event.target.value);
+          }}
+          autoFocus
+        />
+      </UserPasswordWrapper>
+      <p
+        style={{
+          color: "#868e96",
+          fontSize: ".875rem",
+          paddingTop: "15px",
+          paddingLeft: "30px",
+          paddingBottom: "15px",
+          borderBottom: "1px solid grey",
+        }}
+      >
+        비밀번호를 입력해주세요.
+      </p>
+      <div style={{ display: "flex", justifyContent: "start" }}>
+        <EditButtons>
+          <button
+            className="save"
+            name="save"
+            onClick={() => {
+              const data = {
+                member_id: member_id,
+                new_password: password,
+                new_nickname: updateNickName,
+                updated_stacks: personName,
+              };
+              const formData = new FormData();
+              formData.append(
+                "request",
+                new Blob([JSON.stringify(data)], {
+                  type: "application/json",
+                })
+              );
+              dispatch(updateAsync(formData));
+            }}
+          >
+            완료
+          </button>
+          <button className="user_out" name="userOut">
+            회원탈퇴
+          </button>
+        </EditButtons>
+      </div>
     </div>
   );
 };
@@ -146,6 +342,82 @@ const ImageDelete = styled.button`
     color: white;
     transition: ease-out 0.2s;
     background-position: 99% 50%;
+  }
+`;
+const UserNameWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  h3 {
+    width: 20rem;
+  }
+  input {
+    border-radius: 4px;
+    border: 1px solid #ccc;
+    box-sizing: border-box;
+    width: 18rem;
+    min-height: 3rem;
+    padding: 1rem;
+  }
+`;
+const UserPasswordWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  h3 {
+    width: 20rem;
+  }
+  input {
+    border-radius: 4px;
+    border: 1px solid #ccc;
+    box-sizing: border-box;
+    width: 18rem;
+    min-height: 3rem;
+    padding: 1rem;
+  }
+`;
+const TechWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  h3 {
+    width: 20rem;
+    margin: 30px;
+    display: block;
+    font-size: 1.17em;
+    margin-block-start: 1em;
+    margin-block-end: 1em;
+    margin-inline-start: 0px;
+    margin-inline-end: 0px;
+    font-weight: bold;
+    padding: 20px;
+  }
+`;
+const EditButtons = styled.div`
+  width: 800px;
+  margin: 0 auto;
+  display: flex;
+  align-items: center;
+  margin-top: 1.5rem;
+  .save {
+    background-color: #262626;
+    font-weight: 700;
+    color: #fff;
+    border-radius: 4px;
+    width: 100px;
+    margin-right: 10px;
+    height: 2rem;
+    font-size: 1rem;
+    border: 0;
+    cursor: pointer;
+  }
+  .user_out {
+    background: #ff3217;
+    color: #fff;
+    border-radius: 4px;
+    width: 100px;
+    margin-right: 10px;
+    height: 2rem;
+    font-size: 1rem;
+    border: 0;
+    cursor: pointer;
   }
 `;
 export default UserImg;
