--- src/components/InformationEdit/UserStack.js
@@ -1,126 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-import { Theme, useTheme } from "@mui/material/styles";
-import Box from "@mui/material/Box";
-import OutlinedInput from "@mui/material/OutlinedInput";
-import InputLabel from "@mui/material/InputLabel";
-import MenuItem from "@mui/material/MenuItem";
-import FormControl from "@mui/material/FormControl";
-import Select, { SelectChangeEvent } from "@mui/material/Select";
-import Chip from "@mui/material/Chip";
-// 내정보 수정 화면에서 기술스텍 부분
-const ITEM_HEIGHT = 48;
-const ITEM_PADDING_TOP = 8;
-const MenuProps = {
-  PaperProps: {
-    style: {
-      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-      width: 250,
-    },
-  },
-};
-
-const names = [
-  "JAVA",
-  "JavaScript",
-  "React",
-  "Vue",
-  "Python",
-  "Spring",
-  "TypeScript",
-  "Angular",
-  "C++",
-  "Kotlin",
-  "Svelte",
-];
-
-function getStyles(name, personName, theme) {
-  return {
-    fontWeight:
-      personName.indexOf(name) === -1
-        ? theme.typography.fontWeightRegular
-        : theme.typography.fontWeightMedium,
-  };
-}
-
-const UserStack = () => {
-  const theme = useTheme();
-  const [personName, setPersonName] = React.useState([]);
-
-  const handleChange = event => {
-    const {
-      target: { value },
-    } = event;
-    setPersonName(
-      // On autofill we get a stringified value.
-      typeof value === "string" ? value.split(",") : value
-    );
-  };
-
-  return (
-    <>
-      <TechWrapper>
-        <h3>관심 기술 태그</h3>
-        <FormControl sx={{ m: 1, width: 300 }}>
-          <InputLabel id="demo-multiple-chip-label">TechStack</InputLabel>
-          <Select
-            labelId="demo-multiple-chip-label"
-            id="demo-multiple-chip"
-            multiple
-            value={personName}
-            onChange={handleChange}
-            input={<OutlinedInput id="select-multiple-chip" label="Chip" />}
-            renderValue={selected => (
-              <Box sx={{ display: "flex", flexWrap: "wrap", gap: 0.5 }}>
-                {selected.map(value => (
-                  <Chip key={value} label={value} />
-                ))}
-              </Box>
-            )}
-            MenuProps={MenuProps}
-          >
-            {names.map(name => (
-              <MenuItem
-                key={name}
-                value={name}
-                style={getStyles(name, personName, theme)}
-              >
-                {name}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
-      </TechWrapper>
-      <p
-        style={{
-          color: "#868e96",
-          fontSize: ".875rem",
-          paddingTop: "15px",
-          paddingLeft: "30px",
-          paddingBottom: "15px",
-          borderBottom: "1px solid grey",
-        }}
-      >
-        관심 있는 기술 태그를 입력해주세요.
-      </p>
-    </>
-  );
-};
-
-const TechWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  h3 {
-    width: 20rem;
-    margin: 30px;
-    display: block;
-    font-size: 1.17em;
-    margin-block-start: 1em;
-    margin-block-end: 1em;
-    margin-inline-start: 0px;
-    margin-inline-end: 0px;
-    font-weight: bold;
-    padding: 20px;
-  }
-`;
-export default UserStack;
